tests/flet-app/pyproject.toml
[project]
name = "flet-app"
version = "0.1.0"
license = "MIT"
dependencies = ["flet==0.25.*", "flet-easy==0.2.8", "tortoise-orm==0.22.2"]

[[project.authors]]
name = "Daxexs"
email = "Daxexs@test.com"

[tool.flet]
# org name in reverse domain name notation, e.g. "com.mycompany".
# Combined with project.name to build bundle ID for iOS and Android apps
org = "com.mycompany"

# project display name that is used as an app title on Android and iOS home screens,
# shown in window titles and about app dialogs on desktop.
product = "flet-app"

# company name to display in about app dialogs
company = "Flet"

# copyright text to display in about app dialogs
copyright = "Copyright (C) 2024 by Flet"

[tool.flet.app]
path = "src"
